update semua dan menghapus display.js
