--- app/config/display.js
@@ -1,9 +1,0 @@
-function err(res,msg = "default"){
-    if (msg == "default"){
-        res.json({"message":"server connection error"})
-    }else{
-        res.json({"message":msg})
-    }
-}
-
-module.exports = { err }
